座標変換のパラメータの計算をファイルに切り出す
